refactor: create composables for core game logic and apply to additionGame
