+++ src/config/gameConfigs.js
@@ -1,0 +1,12 @@
+export const gameConfigs = {
+  addition: {
+    title: 'Animal Addition',
+    description: 'Add up the animals!',
+    bgColor: '#FFBCEE',
+    iconSrc: '/assets/gameImages/buttons/gameButtons/addition.svg',
+    introAudio: '/animalAddition/additionintro.mp3',
+    dbFile: 'additionDb.json',
+    dbKey: 'AdditionGame',
+    category: 'math',
+  },
+};
